+++ src/components/Message/MessageSimple.js
@@ -8,7 +8,10 @@ import { Attachment as DefaultAttachment } from '../Attachment';
 import { Avatar as DefaultAvatar } from '../Avatar';
 import { MML } from '../MML';
 import { Modal } from '../Modal';
-import { MessageInput, EditMessageForm } from '../MessageInput';
+import {
+  MessageInput,
+  EditMessageForm as DefaultEditMessageForm,
+} from '../MessageInput';
 import { Tooltip } from '../Tooltip';
 import { LoadingIndicator } from '../Loading';
 import {
@@ -47,6 +50,7 @@ const MessageSimple = (props) => {
   const {
     clearEditingState,
     editing,
+    EditMessageInput = DefaultEditMessageForm,
     message,
     threadList,
     formatDate,
@@ -104,7 +108,7 @@ const MessageSimple = (props) => {
       {editing && (
         <Modal open={editing} onClose={clearEditingState}>
           <MessageInput
-            Input={EditMessageForm}
+            Input={EditMessageInput}
             message={message}
             clearEditingState={clearEditingState}
             updateMessage={updateMessage}
@@ -121,7 +125,12 @@ const MessageSimple = (props) => {
 						str-chat__message--${message.status}
 						${message.text ? 'str-chat__message--has-text' : 'has-no-text'}
 						${hasAttachment ? 'str-chat__message--has-attachment' : ''}
-						${hasReactions && isReactionEnabled ? 'str-chat__message--with-reactions' : ''}
+            ${
+              hasReactions && isReactionEnabled
+                ? 'str-chat__message--with-reactions'
+                : ''
+            }
+            ${message?.pinned ? 'pinned-message' : ''}
 					`.trim()}
           ref={messageWrapperRef}
         >
@@ -340,6 +349,12 @@ MessageSimple.propTypes = {
    * */
   Avatar: /** @type {PropTypes.Validator<React.ElementType<import('types').AvatarProps>>} */ (PropTypes.elementType),
   /**
+   * Custom UI component to override default edit message input
+   *
+   * Defaults to and accepts same props as: [EditMessageForm](https://github.com/GetStream/stream-chat-react/blob/master/src/components/MessageInput/EditMessageForm.js)
+   * */
+  EditMessageInput: /** @type {PropTypes.Validator<React.FC<import("types").MessageInputProps>>} */ (PropTypes.elementType),
+  /**
    * @deprecated Its not recommended to use this anymore. All the methods in this HOC are provided explicitly.
    *
    * The higher order message component, most logic is delegated to this component
@@ -363,7 +378,7 @@ MessageSimple.propTypes = {
   /** If component is in thread list */
   threadList: PropTypes.bool,
   /**
-   * Function to open thread on current messxage
+   * Function to open thread on current message
    * @deprecated The component now relies on the useThreadHandler custom hook
    * You can customize the behaviour for your thread handler on the <Channel> component instead.
    */
@@ -375,7 +390,7 @@ MessageSimple.propTypes = {
   /** Returns true if message belongs to current user */
   isMyMessage: PropTypes.func,
   /**
-   * Returns all allowed actions on message by current user e.g., [edit, delete, flag, mute]
+   * Returns all allowed actions on message by current user e.g., ['edit', 'delete', 'flag', 'mute', 'react', 'reply']
    * Please check [Message](https://github.com/GetStream/stream-chat-react/blob/master/src/components/Message.js) component for default implementation.
    * */
   getMessageActions: PropTypes.func.isRequired,