+++ src/components/Message/__tests__/MessageTeam.test.js
@@ -1,6 +1,6 @@
 /* eslint-disable sonarjs/no-duplicate-string */
 import React from 'react';
-import { cleanup, render, fireEvent } from '@testing-library/react';
+import { cleanup, render, fireEvent, waitFor } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import {
   generateChannel,
@@ -10,7 +10,7 @@ import {
   generateReaction,
 } from 'mock-builders';
 
-import { ChannelContext } from '../../../context';
+import { ChannelContext, TranslationContext } from '../../../context';
 import MessageTeam from '../MessageTeam';
 import { Avatar as AvatarMock } from '../../Avatar';
 import { MML as MMLMock } from '../../MML';
@@ -42,9 +42,19 @@ async function renderMessageTeam(
 ) {
   const channel = generateChannel({ getConfig: () => channelConfig });
   const client = await getTestClientWithUser(alice);
+  const customDateTimeParser = jest.fn(() => ({ format: jest.fn() }));
+
   return render(
     <ChannelContext.Provider value={{ client, channel, t: (key) => key }}>
-      <MessageTeam message={message} typing={false} {...props} />
+      <TranslationContext.Provider
+        value={{
+          t: (key) => key,
+          tDateTimeParser: customDateTimeParser,
+          userLanguage: 'en',
+        }}
+      >
+        <MessageTeam message={message} typing={false} {...props} />
+      </TranslationContext.Provider>
     </ChannelContext.Provider>,
   );
 }
@@ -160,6 +170,61 @@ describe('<MessageTeam />', () => {
     expect(getByTestId('custom-avatar')).toBeInTheDocument();
   });
 
+  it('should render pin indicator when pinned is true', async () => {
+    const message = generateAliceMessage({ pinned: true });
+    const CustomPinIndicator = () => (
+      <div data-testid="pin-indicator">Pin Indicator</div>
+    );
+
+    const { getByTestId } = await renderMessageTeam(message, {
+      PinIndicator: CustomPinIndicator,
+    });
+
+    await waitFor(() => {
+      expect(getByTestId('pin-indicator')).toBeInTheDocument();
+    });
+  });
+
+  it('should not render pin indicator when pinned is false', async () => {
+    const message = generateAliceMessage({ pinned: false });
+    const CustomPinIndicator = () => (
+      <div data-testid="pin-indicator">Pin Indicator</div>
+    );
+
+    const { queryAllByTestId } = await renderMessageTeam(message, {
+      PinIndicator: CustomPinIndicator,
+    });
+
+    await waitFor(() => {
+      expect(queryAllByTestId('pin-indicator')).toHaveLength(0);
+    });
+  });
+
+  it('should render custom edit message input component when one is given', async () => {
+    const message = generateAliceMessage();
+    const updateMessage = jest.fn();
+    const clearEditingState = jest.fn();
+
+    const CustomEditMessageInput = () => <div>Edit Input</div>;
+
+    await renderMessageTeam(message, {
+      clearEditingState,
+      editing: true,
+      updateMessage,
+      EditMessageInput: CustomEditMessageInput,
+    });
+
+    expect(MessageInputMock).toHaveBeenCalledWith(
+      expect.objectContaining({
+        clearEditingState,
+        message,
+        Input: CustomEditMessageInput,
+        updateMessage,
+      }),
+      {},
+    );
+  });
+
   it('should render message input when in edit mode', async () => {
     const message = generateAliceMessage();
     const updateMessage = jest.fn();
@@ -237,6 +302,17 @@ describe('<MessageTeam />', () => {
       },
       {},
     );
+  });
+
+  it('should display text in users set language', async () => {
+    const message = generateAliceMessage({
+      i18n: { fr_text: 'bonjour', en_text: 'hello', language: 'fr' },
+      text: 'bonjour',
+    });
+
+    const { getByText } = await renderMessageTeam(message);
+
+    expect(getByText('hello')).toBeInTheDocument();
   });
 
   it('should place a spacer when message is not the first message on a thread and group style is not top or single', async () => {